requirements.txt
yfinance>=0.2.36
feedparser>=6.0.11
groq>=1.0.0
resend>=2.0.0
python-dotenv>=1.0.1
jinja2>=3.1.4
